requirements.txt
pyyaml>=6.0.1
graphviz>=0.20.1
mkdocs>=1.5.3
pytest>=7.4.3
python-frontmatter>=1.0.0
click>=8.1.7
aiohttp>=3.9.1
asyncio>=3.4.3
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.7
python-multipart>=0.0.6 